requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
pydantic
langchain-groq
langchain-openai
python-dotenv